dernières modifications avant soutenance

+++ back/controllers/com.js
@@ -1,4 +1,5 @@
 const Com = require('../models/com');
+const User = require('../models/user')
 const fs = require('fs');
 
 //Logique métier pour créer un nouveau com
@@ -25,8 +26,9 @@ exports.createCom = (req, res, next) => {
     }
   
     com.save()
-    .then(() => { console.log(comObject);
-        res.status(201).json({message: 'Commentaire enregistré !'})})
+    .then(() => { 
+        res.status(201).json({message: 'Commentaire enregistré !'})
+    })
     .catch(error => { 
         res.status(400).json( { error })
     })
@@ -35,9 +37,7 @@ exports.createCom = (req, res, next) => {
 //Logique métier pour modifier un com
 
 exports.modifyCom = (req, res, next) => {
-    console.log(req.body);
     const comObject = req.file ? {
-        //...JSON.parse(req.body.text),
         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     } : { ...req.body };
 
@@ -45,23 +45,27 @@ exports.modifyCom = (req, res, next) => {
     Com.findOne({_id: req.params.id})
 
         .then((com) => {
-            if (com.userId != req.auth.userId) {
-                res.status(403).json({ message : 'Unauthorized request'});
-            } else {
-
-                //Gestion de l'image lors de la modification d'un com
-                if (req.file) {
-                    const filename = com.imageUrl.split('/images/')[1];
-                    if (fs.existsSync(`images/${filename}`)) {
-                        fs.unlinkSync(`images/${filename}`);
-                    } 
-                }
-
-                Com.updateOne({ _id: req.params.id}, { ...comObject, _id: req.params.id})
-                .then(() => res.status(200).json({message : 'Objet modifié!'}))
-                .catch(error => res.status(401).json({ error }));
-            }
+            User.findById(req.auth.userId)
+                .then((user) => {
+                    if (!user.isAdmin && com.userId != req.auth.userId) {
+                        res.status(403).json({ message : 'Unauthorized request'});
+                    } else {
+        
+                        //Gestion de l'image lors de la modification d'un com
+                        if (req.file) {
+                            const filename = com.imageUrl.split('/images/')[1];
+                            if (fs.existsSync(`images/${filename}`)) {
+                                fs.unlinkSync(`images/${filename}`);
+                            } 
+                        }
+        
+                        Com.updateOne({ _id: req.params.id}, { ...comObject, _id: req.params.id})
+                        .then(() => res.status(200).json({message : 'Objet modifié!'}))
+                        .catch(error => res.status(401).json({ error }));
+                    }
+                });
         })
+            
         .catch((error) => {
             res.status(400).json({ error });
         });
@@ -73,23 +77,27 @@ exports.modifyCom = (req, res, next) => {
 exports.deleteCom = (req, res, next) => {
 Com.findOne({ _id: req.params.id})
     .then(com => {
-        if (com.userId != req.auth.userId) {
-            res.status(403).json({message: 'Unauthorized request'});
-        } else {
-            if(com.imageUrl) {
-            const filename = com.imageUrl.split('/images/')[1];
-            fs.unlink(`images/${filename}`, () => {
-                Com.deleteOne({_id: req.params.id})
-                    .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
-                    .catch(error => res.status(401).json({ error }));
+        User.findById(req.auth.userId)
+            .then((user) => {
+                if (!user.isAdmin && com.userId != req.auth.userId) {
+                    res.status(403).json({message: 'Unauthorized request'});
+                } else {
+                    if(com.imageUrl) {
+                    const filename = com.imageUrl.split('/images/')[1];
+                    fs.unlink(`images/${filename}`, () => {
+                        Com.deleteOne({_id: req.params.id})
+                            .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                            .catch(error => res.status(401).json({ error }));
+                    });
+                    } else {
+                        Com.deleteOne({_id: req.params.id})
+                            .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                            .catch(error => res.status(401).json({ error }));
+                    }
+                }
             });
-            } else {
-                Com.deleteOne({_id: req.params.id})
-                    .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
-                    .catch(error => res.status(401).json({ error }));
-            }
-        }
     })
+
     .catch( error => {
         res.status(500).json({ error });
     });
@@ -99,7 +107,6 @@ Com.findOne({ _id: req.params.id})
 
 exports.getAllComs = (req, res, next) => {
 Com.find()
-    .sort('_id')
     .then(coms => res.status(200).json(coms))
     .catch(error => res.status(400).json({ error }));
 }
