+++ back/controllers/post.js
@@ -1,4 +1,6 @@
 const Post = require("../models/post");
+const User = require("../models/user");
+const Com = require("../models/com")
 const fs = require('fs');
 
 //Logique métier pour créer un nouveau post
@@ -24,8 +26,9 @@ exports.createPost = (req, res, next) => {
     }
     
     post.save()
-    .then(() => { console.log(postObject);
-        res.status(201).json({message: 'Objet enregistré !'})})
+    .then(() => { 
+        res.status(201).json({message: 'Objet enregistré !'})
+    })
     .catch(error => { 
         res.status(400).json( { error })
     })
@@ -36,34 +39,35 @@ exports.createPost = (req, res, next) => {
 exports.modifyPost = (req, res, next) => {
     
     const postObject = req.file ? {
-        /*...JSON.parse(req.body.text),*/
         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     } : { ...req.body };
     
     delete postObject._userId;
     Post.findOne({_id: req.params.id})
-
         .then((post) => {
-            if (post.userId != req.auth.userId) {
-                res.status(403).json({ message : 'Unauthorized request'});
-            } else {
-
-                //Gestion de l'image lors de la modification d'un post
-                if (req.file) {
-                    const filename = post.imageUrl.split('/images/')[1];
-                    if (fs.existsSync(`images/${filename}`)) {
-                        fs.unlinkSync(`images/${filename}`);
-                    } 
+            User.findById(req.auth.userId)
+            .then((user) => {
+                if (!user.isAdmin && post.userId != req.auth.userId) {
+                    res.status(403).json({ message : 'Unauthorized request'});
+                } else {
+    
+                    //Gestion de l'image lors de la modification d'un post
+                    if (req.file) {
+                        const filename = post.imageUrl.split('/images/')[1];
+                        if (fs.existsSync(`images/${filename}`)) {
+                            fs.unlinkSync(`images/${filename}`);
+                        } 
+                    }
+    
+                    Post.updateOne({ _id: req.params.id}, {
+                         ...postObject,
+                         updatedAt: Date.now(),
+                         _id: req.params.id
+                        })
+                    .then( res.status(200).json({message : 'Objet modifié!'}))
+                    .catch(error => res.status(401).json({ error }));
                 }
-
-                Post.updateOne({ _id: req.params.id}, {
-                     ...postObject,
-                     updatedAt: Date.now(),
-                     _id: req.params.id
-                    })
-                .then( res.status(200).json({message : 'Objet modifié!'}))
-                .catch(error => res.status(401).json({ error }));
-            }
+            });
         })
         .catch((error) => {
             res.status(400).json({ error });
@@ -75,28 +79,52 @@ exports.modifyPost = (req, res, next) => {
 
 exports.deletePost = (req, res, next) => {
 Post.findOne({ _id: req.params.id})
-    .then(post => {
-        if (post.userId != req.auth.userId) {
-            res.status(403).json({message: 'Unauthorized request'});
-        } else {
-            if (post.imageUrl) {
-                const filename = post.imageUrl.split('/images/')[1];
-                fs.unlink(`images/${filename}`, () => {
-                    Post.deleteOne({_id: req.params.id})
-                        .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
-                        .catch(error => res.status(401).json({ error }));
+    
+        .then((post) => {
+            User.findById(req.auth.userId)
+                .then((user) => {
+                    
+                    if (!user.isAdmin && post.userId != req.auth.userId) {
+                        res.status(403).json({message: 'Unauthorized request'});
+                    } else {
+                        deleteAllComOfPost(post.id);
+                        if (post.imageUrl) {
+                            const filename = post.imageUrl.split('/images/')[1];
+                            fs.unlink(`images/${filename}`, () => {
+                                Post.deleteOne({_id: req.params.id})
+                                    .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                                    .catch(error => res.status(401).json({ error }));
+                            });
+                        } else {
+                            Post.deleteOne({_id: req.params.id})
+                            .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                            .catch(error => res.status(401).json({ error }));
+                        }
+                    }
                 });
-            } else {
-                Post.deleteOne({_id: req.params.id})
-                .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
-                .catch(error => res.status(401).json({ error }));
-            }
-        }
-    })
-    .catch( error => {
-        res.status(500).json({ error });
-    });
+            })
+        
+
+        .catch( error => {
+            res.status(500).json({ error });
+        });
 };
+
+function deleteAllComOfPost(postId) {
+    Com.find({postId})
+      .then((comsPost) => {
+        comsPost.forEach(com => {
+          if (com.imageUrl) {
+            const filename = com.imageUrl.split('/images/')[1];
+            fs.unlink(`images/${filename}`, () => {
+              Com.deleteOne({_id: com._id});
+            });
+          } else {
+            Com.deleteOne({_id: com._id});
+          }
+        });
+      })
+  }
 
 //Logique métier pour récupérer tous les posts
 
@@ -149,12 +177,10 @@ exports.likePost = (req, res, next) => {
 
         if (like === 0) {
             if (indexLike === -1) {
-                //ici le probleme, indexDislike n'est pas l'index du user mais le code envoyé par le front (1, 0 ou -1)
                 post.usersDisliked.splice(indexDislike, 1)
                 post.dislikes = post.usersDisliked.length
             }
             else {
-                //ici aussi le probleme
                 post.usersLiked.splice(indexLike, 1)
                 post.likes = post.usersLiked.length
             }
